fix test/ directory
I would like to eventually have a working testsuite target in the main
Makefile, but a good first step is to make sure the various tests are
usable in the new repository. This commit fixes the build system for
the old test/ tests (more recent tests rather go in testsuite/), and
fixes the tests cleanup logic to leave the repository in a clean state
after a correct run. To try it:

  cd test
  sh testsuite.sh

(There is currently no checking that tests indeed pass, maybe later.)