consider files starting with '_' as source files
This was requested by Daniel Bünzli on the caml-list:
     Subject: [Caml-list] ocamlbuild, build a source that starts with _
     Date: Tue, 9 Feb 2016 14:35:06 +0100
     https://sympa.inria.fr/sympa/arc/caml-list/2016-02/msg00033.html

It does not make sense for .ml files, as .ml basenames must be valid
OCaml modules, so a .ml file cannot start with _ anyway, but it does
make sense for C files for example.

This commit simply removes the special logic that would disregard
_-starting files when slurping the directory. I suppose this logic was
meant to prevent _oasis, _tags etc. from being considered in build
rules, but I don't see any negative side-effect in undoing it.