make sure that -just-plugin always stops after the plugin-build phase
Previously -just-plugin would only be checked at the end of plugin
build, and plugin build would only run if there is a plugin. This
means that -just-plugin would be ignored if there was no
myocamlbuild.ml: (ocamlbuild -just-plugin foo.byte) would build the
target or not depending on whether a plugin source was present.