Makefile: separate clean and distclean targets
`clean` only cleans the temporary targets, but leaves in the
repository the ocamlbuild.{byte,native} executables that are useful to
play with the build result. (All *.cm* files are cleaned, however, so
there is not enough to install ocamlbuild.)

`distclean` cleans everything, including in particular the
configuration files used during the build, and the final executables.

See <https://www.gnu.org/prep/standards/html_node/Standard-Targets.html>
for example for a description of those standard Makefile targets.