Defer to the compiler in choosing how to build a shared library.
This commit performs two changes:
  * When building a shared library through -output-obj, ocamlbuild
    will no longer try and second-guess the compiler with regards
    to how to build it, by passing a -cclib -shared flag.
    This was harmless on Windows and *nix, but produced errors
    on macOS; see #124 and ocaml/ocaml#988.
  * The documentation no longer mentions the .dylib extension,
    which was not used for anything by ocaml or ocamlbuild for
    a very long time (forever).

Due to ocaml/ocaml#988, there is a subtle difference in behavior
when ocamlbuild is used on 4.04 and previous versions versus
4.05:
  * On 4.04 and earlier versions, it was not possible to build
    an x.{byte,native}.so on macOS as the linker flags conflicted.
    After this commit, this is possible and produces a Mach-O
    bundle.
  * On 4.05 and later versions, passing -cclib -shared is a no-op.

Since it is not unusual that x.{byte,native}.so is explicitly loaded
dynamically, I have decided to remove the -cclib -shared flag
instead of leaving it as a no-op, to let people using new ocamlbuild
with old OCaml compilers take advantage of that.

The output_shared tag is left for backwards compatibility with
myocamlbuild.ml code that adds flags through it.