Let autolink do its job when linking .cmxs files.
Before the recent .cmxs refactor, i.e. 89888f2c4...f865497a8, .cmxs
files were built from .cmx and .o files. After commit f865497a8,
.cmxs files are built from .cmxa and .a files, which is wrong,
because .cmxa files already contain the autolink instructions
necessary to link the .a file; and on top of that, no -I flag
was provided when linking the .cmxs file, which resulted in
the linker being unable to follow the autolink instructions.

This caused failures in the wild when ocamlbuild 0.10.0 got released:
https://github.com/ocaml/opam-repository/pull/8157.

After this commit, .a files are no longer explicitly passed
when building a .cmxs file. Additionally, for every linked file,
a -I flag is passed, with the directory containing the file
being linked.