Make the testsuite automatically discover and use the build tree.
Before this commit, it was very easy to accidentally use artifacts
from the installed ocamlbuild instance. In fact, it was impossible
to test linking of plugins properly, because plugins were always
linked with the system ocamlbuildlib.cm{x,}a.

After this commit, the testsuite does not need any user action
to use the correct ocamlbuild executable or ocamlbuildlib.cm{x,}a.