Provide `-project` command line option to avoid recursive traversal hint.
The current heuristic (Options.ocamlbuild_project_heuristic) checks for either
`_tags` or `myocamlbuild.ml` to be present, or "-recursive" being passed to
ocamlbuild.  This works fine for many applications, but e.g. for mirage
unikernels it does not (they are in a flat directory, subdirectories may be
present containing data which should not be traversed).

At the moment, mirage creates an empty `myocamlbuild.ml` before calling
ocamlbuild to avoid this (for MirageOS users misleading) hint.